+++ docsite/docs/more/release_notes.md
@@ -3,6 +3,7 @@
 ## 0.5.6
 
 - **Added a [Yarn plugin](../plugins_reference/yarn/yarn_overview) to handle Yarn services (long running jobs) lifecycle.**
+- Storm topologies ACTIVE status is now displayed in upper case, for better visibility.
 
 ## 0.5.5